+++ inst/examples/get_gbif_help.R
@@ -1,16 +1,19 @@
 \donttest{
-# Download all worldwide observations of Ailuropoda melanoleuca, with:
+# Download all worldwide observations of the great panda, with:
 # - 100km grain
 # - after 1990
-# - keeping duplicates and by
+# - keeping duplicates
 # - adding the name of the person who collected the panda records
-obs.am <- get_gbif(
+obs_am <- get_gbif(
        sp_name = "Ailuropoda melanoleuca",
-       grain = 100 ,
+       grain = 100,
        duplicates = TRUE,
        time_period = c(1990,3000),
        add_infos = c("recordedBy","issue")
 )
+
+# Guard against an unavailable remote source
+if (gbif_have(obs_am)) {
 
 # Extract borders
 countries <- terra::vect(
@@ -20,10 +23,11 @@ countries <- terra::vect(
 # Plot
 terra::plot(countries, col = "#bcbddc")
 graphics::points(
-       obs.am[,c("decimalLongitude","decimalLatitude")],
+       obs_am[,c("decimalLongitude","decimalLatitude")],
        pch = 20,
        col = "#238b4550",
        cex = 4
 )
 
+}
 }