+++ inst/examples/make_tiles_help.R
@@ -1,9 +1,9 @@
 # Load the European Alps Extent
-shp.path <- paste0(
+shp_path <- paste0(
 	system.file(package = "gbif.range"), 
 	"/extdata/shp_lonlat.shp"
 )
-shp.lonlat <- terra::vect(shp.path)
+shp_lonlat <- terra::vect(shp_path)
  
 # Apply the function to divide the extent in ~20 fragments
-mt <- make_tiles(geo = shp.lonlat, ntiles = 20, sext = TRUE)
+mt <- make_tiles(geo = shp_lonlat, ntiles = 20, sext = TRUE)