+++ NEWS.md
@@ -1,9 +1,18 @@
+# gbif.range 1.9.0
+* Added merge_range()
+* Improved speed of get_range()
+* Corrected cluster convergence beug in get_range()
+* Corrected column error when get_status set with level = "all"
+* CRAN compliance v2
+* Corrected documentation
+
 # gbif.range 1.8.0
 * Move manuscript plots in a dedicated vignette
 * Updated the examples and vignettes
 * Lower the size and time length of the examples
 * Debugging
 * Documentation updated
+* CRAN compliance v1
 
 # gbif.range 1.7.1
 * Updated documentation